# planung/restapi/api/src/main/java/org/openapitools/api/AreaApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.1.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.Area;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2023-11-19T14:12:38.522563+01:00[Europe/Berlin]")
@Validated
@Tag(name = "Area", description = "Access to Area data")
public interface AreaApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /wecker/areas/ : Add Area with ID
     * Add Area with ID
     *
     * @param area Add Area (optional)
     * @return Successful operation (status code 200)
     *         or Invalid ID supplied (status code 400)
     *         or Validation exception (status code 405)
     */
    @Operation(
        operationId = "addArea",
        summary = "Add Area with ID",
        description = "Add Area with ID",
        tags = { "Area" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Area.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid ID supplied"),
            @ApiResponse(responseCode = "405", description = "Validation exception")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/wecker/areas/",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Area> addArea(
        @Parameter(name = "Area", description = "Add Area") @Valid @RequestBody(required = false) Area area
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"name\" : \"Bett\", \"id\" : 123 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /wecker/areas/ : Get all Areas
     * Get all Areas
     *
     * @return Successful operation (status code 200)
     */
    @Operation(
        operationId = "getAllAreas",
        summary = "Get all Areas",
        description = "Get all Areas",
        tags = { "Area" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Area.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/wecker/areas/",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Area>> getAllAreas(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"name\" : \"Bett\", \"id\" : 123 }, { \"name\" : \"Bett\", \"id\" : 123 } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /wecker/areas/{areaId} : Get Area via ID
     * Get Area via ID
     *
     * @param areaId ID of Area (required)
     * @return Successful operation (status code 200)
     *         or Invalid ID supplied (status code 400)
     *         or Area not found (status code 404)
     *         or Validation exception (status code 405)
     */
    @Operation(
        operationId = "getAreaById",
        summary = "Get Area via ID",
        description = "Get Area via ID",
        tags = { "Area" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Area.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid ID supplied"),
            @ApiResponse(responseCode = "404", description = "Area not found"),
            @ApiResponse(responseCode = "405", description = "Validation exception")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/wecker/areas/{areaId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Area> getAreaById(
        @Parameter(name = "areaId", description = "ID of Area", required = true, in = ParameterIn.PATH) @PathVariable("areaId") Long areaId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"name\" : \"Bett\", \"id\" : 123 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /wecker/areas/{areaId} : Update Area with ID
     * Update Area with ID
     *
     * @param areaId ID of Area (required)
     * @param area Update an existent Area (optional)
     * @return Successful operation (status code 200)
     *         or Invalid ID supplied (status code 400)
     *         or Area not found (status code 404)
     *         or Validation exception (status code 405)
     */
    @Operation(
        operationId = "updateAreaViaId",
        summary = "Update Area with ID",
        description = "Update Area with ID",
        tags = { "Area" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Area.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid ID supplied"),
            @ApiResponse(responseCode = "404", description = "Area not found"),
            @ApiResponse(responseCode = "405", description = "Validation exception")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/wecker/areas/{areaId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Area> updateAreaViaId(
        @Parameter(name = "areaId", description = "ID of Area", required = true, in = ParameterIn.PATH) @PathVariable("areaId") Long areaId,
        @Parameter(name = "Area", description = "Update an existent Area") @Valid @RequestBody(required = false) Area area
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"name\" : \"Bett\", \"id\" : 123 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}