# planung/restapi/api/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: Dies ist eine Schnittstellenbeschreibung für die Wecker-REST-API
  termsOfService: http://swagger.io/terms/
  title: Wecker Rest API - OpenAPI 3.0
  version: 3.1.0
servers:
- url: localhost/api/v1
...
paths:
...
  /wecker/areas/:
    get:
      description: Get all Areas
      operationId: getAllAreas
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Area'
                type: array
          description: Successful operation
      summary: Get all Areas
      tags:
      - Area
      x-accepts: application/json
      x-tags:
      - tag: Area
    post:
      description: Add Area with ID
      operationId: addArea
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Area'
        description: Add Area
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Area'
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "405":
          description: Validation exception
      summary: Add Area with ID
      tags:
      - Area
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Area
  /wecker/areas/{areaId}:
    get:
      description: Get Area via ID
      operationId: getAreaById
      parameters:
      - description: ID of Area
        explode: false
        in: path
        name: areaId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Area'
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Area not found
        "405":
          description: Validation exception
      summary: Get Area via ID
      tags:
      - Area
      x-accepts: application/json
      x-tags:
      - tag: Area
    put:
      description: Update Area with ID
      operationId: updateAreaViaId
      parameters:
      - description: ID of Area
        explode: false
        in: path
        name: areaId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Area'
        description: Update an existent Area
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Area'
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Area not found
        "405":
          description: Validation exception
      summary: Update Area with ID
      tags:
      - Area
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Area
...
components:
...
  schemas:
    Area:
      example:
        name: Bett
        id: 123
      properties:
        id:
          example: 123
          format: int64
          type: integer
        name:
          example: Bett
          type: string
      type: object
...